% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_all_close}
\alias{are_all_close}
\title{Test two vectors are close}
\usage{
are_all_close(v, w, abs_tol = 1e-06, rel_tol = 1e-06)
}
\arguments{
\item{v, }{vector.}

\item{w, }{vector of the same length as v.}

\item{abs_tol, }{number, tolerance for the absolute error.}

\item{rel_tol, }{number, tolerance for the relative error.}
}
\value{
T&F for the testing.
}
\description{
Test two vectors are close in the sense of both relative error and absolute error.
}
