% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{coef.hglm}
\alias{coef.hglm}
\title{Coefficients of the GLM model}
\usage{
\method{coef}{hglm}(hglm_out)
}
\arguments{
\item{hglm_out, }{output of the GLM model}
}
\value{
coefficient vector
}
\description{
Coefficients of the GLM model
}
