% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Sigmoid (expit) function}
\usage{
expit(x)
}
\arguments{
\item{x}{Numeric vector or matrix.}
}
\value{
Transformed values using the expit function.
}
\description{
Computes the expit (logistic sigmoid) function.
}
\keyword{internal}
