% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{extract_R}
\alias{extract_R}
\title{Extract the R Factor from a QR Decomposition}
\usage{
extract_R(X)
}
\arguments{
\item{X}{A numeric matrix.}
}
\value{
The reordered upper-triangular matrix \eqn{R}.
}
\description{
This function performs a QR decomposition on a given matrix \eqn{X}, extracts the upper-triangular
factor \eqn{R}, and reorders its columns according to the pivoting information. This is useful for
ensuring that \eqn{X \approx Q \times R} holds, even when column pivoting is used.
}
\keyword{internal}
