% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{get_neg_likelihood_functions}
\alias{get_neg_likelihood_functions}
\title{\subsection{--------------choose based on model}{
}}
\usage{
get_neg_likelihood_functions(model)
}
\description{
\subsection{--------------choose based on model}{
}
}
