% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{grad_logit}
\alias{grad_logit}
\title{Gradient for the likelihood under linear model}
\usage{
grad_logit(design, outcome, beta)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{beta}{vector, the linear coefficients}
}
\value{
Gradient for the likelihood under linear model
}
\description{
Calculate the gradient for the logit model
}
