% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{hessian_logistic}
\alias{hessian_logistic}
\title{\subsection{--------------------------- hessian}{
}}
\usage{
hessian_logistic(beta, X, y)
}
\description{
\subsection{--------------------------- hessian}{
}
}
