% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{hiper_glm}
\alias{hiper_glm}
\title{GLM function for high-dimensional data}
\usage{
hiper_glm(X, y, model = "linear", option = list())
}
\arguments{
\item{X}{the design matrix}

\item{y}{the outcome vector/matrix of the model}

\item{model}{choose the type of GLM, currently supporting linear regression or logistic regression}

\item{option}{optional settings for the model, including the choice of MLE solver}
}
\value{
S3 object of the model fitting output
}
\description{
GLM function for high-dimensional data
}
