% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiperglm.R
\name{hiperglm}
\alias{hiperglm}
\title{hiper_glm}
\usage{
hiperglm(design, outcome, optimizer = "pseudo")
}
\arguments{
\item{design}{A numeric matrix representing the linear model design matrix.}

\item{outcome}{A numeric vector representing the response variable observations.}

\item{optimizer}{Character. Optimization method to use. Must be either "pseudo" or "BFGS".
Default is "pseudo".}
}
\value{
A list containing the fitted model components.
}
\description{
Fit a GLM using either pseudo-inverse (OLS) or BFGS optimization.
}
