% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{log_likelihood}
\alias{log_likelihood}
\title{log-likelihood under linear model}
\usage{
log_likelihood(design, outcome, beta, noise_var)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{beta}{vector, the linear coefficients}

\item{noise_var}{number, the variance of the noise}
}
\value{
log-likelihood under linear model
}
\description{
Calculate the log-likelihood
}
