% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{log_likelihood_logit}
\alias{log_likelihood_logit}
\title{log-likelihood under linear model}
\usage{
log_likelihood_logit(design, outcome, beta)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{beta}{vector, the linear coefficients}
}
\value{
log-likelihood under linear model
}
\description{
Calculate the log-likelihood for the logit model
}
