% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{Make_Hessian}
\alias{Make_Hessian}
\title{Hessian matrix calculation}
\usage{
Make_Hessian(design, beta)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{beta}{vector, the linear coefficients.}
}
\value{
The Hessian matrix.
}
\description{
Calculate the Hessian matrix for the current beta.
}
