% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mle_BFGS}
\alias{mle_BFGS}
\title{MLE for linear coefficients}
\usage{
mle_BFGS(design, outcome, fun, grad, ...)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{fun}{function, log-likelihood function taking design, outcome, ... as inputs.}

\item{grad}{function, gradient of the log-likelihood function.}
}
\value{
Estimate of the linear coefficients
}
\description{
Calculate the MLE of linear coefficients using BFGS for general
log-likelihood function and gradient.
}
