% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{mle_pinv}
\alias{mle_pinv}
\title{MLE for linear coefficients}
\usage{
mle_pinv(design, outcome, method = "QR")
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}
}
\value{
Estimate of the linear coefficients
}
\description{
Calculate the MLE of linear coefficients using pseudo-inverse for the linear model
}
