% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{neg_gradient_linear}
\alias{neg_gradient_linear}
\title{Compute gradient of the negative log-likelihood}
\usage{
neg_gradient_linear(beta, X, y)
}
\arguments{
\item{beta}{Coefficients (vector)}

\item{X}{Design matrix}

\item{y}{Response variable}
}
\value{
Gradient vector
}
\description{
Compute gradient of the negative log-likelihood
}
