% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{neg_gradient_logistic}
\alias{neg_gradient_logistic}
\title{Compute gradient of the negative log-likelihood for logistic regression}
\usage{
neg_gradient_logistic(beta, X, y)
}
\arguments{
\item{beta}{Coefficients (vector)}

\item{X}{Design matrix}

\item{y}{Binary response variable (0 or 1)}
}
\value{
Gradient vector
}
\description{
Compute gradient of the negative log-likelihood for logistic regression
}
