% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{neg_log_likelihood_logistic}
\alias{neg_log_likelihood_logistic}
\title{\subsection{------------------------logic model}{
}}
\usage{
neg_log_likelihood_logistic(beta, X, y)
}
\description{
\subsection{------------------------logic model}{
}
}
