% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{Outcome_est}
\alias{Outcome_est}
\title{Response probability}
\usage{
Outcome_est(design, beta)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{beta}{vector, the linear coefficients.}
}
\value{
A probability vector for the response.
}
\description{
Calculate the response probability with the design and coefficient in the logit model.
}
