% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{print.hglm}
\alias{print.hglm}
\title{Print the GLM model fitting results}
\usage{
\method{print}{hglm}(hglm_out)
}
\arguments{
\item{hglm_out, }{output of the GLM model}
}
\value{
summary statistics of the GLM model
}
\description{
Print the GLM model fitting results
}
