% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.hiperglm}
\alias{print.hiperglm}
\title{Summarize Model
Print method for a hiperglm object}
\usage{
\method{print}{hiperglm}(hiperglm_obj)
}
\description{
Summarize Model
Print method for a hiperglm object
}
