% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{Update_step}
\alias{Update_step}
\title{Update step for beta}
\usage{
Update_step(design, outcome, outcome_est, Hessian)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{outcome_est}{vector, the currently estimated probability of response.}

\item{beta}{vector, the linear coefficients.}
}
\value{
A moving step for the linear coefficient beta.
}
\description{
Calculate the update step for the linear coefficient beta using Newton's method.
}
