% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{vcov.hglm}
\alias{vcov.hglm}
\title{Variance-covariance matrix of the GLM model}
\usage{
\method{vcov}{hglm}(hglm_out)
}
\arguments{
\item{hglm_out}{output of the GLM model}
}
\value{
variance-covariance matrix of the GLM model
}
\description{
Variance-covariance matrix of the GLM model
}
