% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.hiperglm}
\alias{vcov.hiperglm}
\title{Extract covariance matrix from a hiperglm object}
\usage{
\method{vcov}{hiperglm}(object, ...)
}
\description{
Extract covariance matrix from a hiperglm object
}
