% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_all_close}
\alias{are_all_close}
\title{Check if two numeric vectors are close to each other}
\usage{
are_all_close(v, w, abs_tol = 1e-06, rel_tol = 1e-06)
}
\arguments{
\item{v}{First numeric vector}

\item{w}{Second numeric vector}

\item{abs_tol}{Absolute tolerance}

\item{rel_tol}{Relative tolerance}
}
\value{
Boolean indicating whether all elements are within tolerance
}
\description{
Check if two numeric vectors are close to each other
}
