% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.hiperglm}
\alias{coef.hiperglm}
\title{Extract model coefficients from a hiper_glm object}
\usage{
\method{coef}{hiperglm}(object, ...)
}
\arguments{
\item{object}{A \code{hiperglm} object}

\item{...}{Additional arguments}
}
\value{
A numeric vector of coefficients
}
\description{
Extract model coefficients from a hiper_glm object
}
