% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_functions.R
\name{fitting_method_pseudo_inverse}
\alias{fitting_method_pseudo_inverse}
\alias{fitting_method_bfgs}
\alias{fitting_method_newton}
\title{Internal Fit Function for hiperglm}
\usage{
fitting_method_pseudo_inverse(design, outcome)

fitting_method_bfgs(design, outcome, option)

fitting_method_newton(design, outcome, option)
}
\description{
Internal Fit Function for hiperglm
}
\section{Pseudo Inverse Method}{


These functions implement the pseudo inverse fitting method.
}

\section{BFGS Fitting Method}{


Implements the BFGS optimization method for fitting.
}

\section{Newton Fitting Method}{


Implements Newton's method for logistic regression.
}

\keyword{internal}
