% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{hessian_logistic}
\alias{hessian_logistic}
\title{\subsection{--------------------------- hessian}{

Compute Hessian matrix for logistic regression
}}
\usage{
hessian_logistic(beta, X, y)
}
\arguments{
\item{beta}{Coefficients (vector)}

\item{X}{Design matrix}

\item{y}{Binary response variable (0 or 1)}
}
\value{
Hessian matrix
}
\description{
\subsection{--------------------------- hessian}{

Compute Hessian matrix for logistic regression
}
}
