% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{hiper_glm}
\alias{hiper_glm}
\title{GLM function for high-dimensional data}
\usage{
hiper_glm(desgin, outcome, model = "linear")
}
\arguments{
\item{outcome}{the outcome vector/matrix of the model}

\item{model}{choose the type of GLM}

\item{design}{the design matrix}
}
\value{
S3 object of the model fitting output
}
\description{
GLM function for high-dimensional data
}
