% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{hiper_glm}
\alias{hiper_glm}
\title{Fit a linear model using either pseudo-inverse or BFGS optimization}
\usage{
hiper_glm(X, y, method = c("pseudo_inverse", "BFGS"))
}
\arguments{
\item{X}{Design matrix (n x p)}

\item{y}{Response variable (n x 1)}

\item{method}{Optimization method: "pseudo_inverse" or "BFGS"}
}
\value{
A \code{hiperglm} object containing estimated coefficients
}
\description{
Fit a linear model using either pseudo-inverse or BFGS optimization
}
