% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{neg_log_likelihood_linear}
\alias{neg_log_likelihood_linear}
\title{\subsection{--------linear model}{
}}
\usage{
neg_log_likelihood_linear(beta, X, y)
}
\description{
\subsection{--------linear model}{
}
}
