% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hiperglm}
\alias{print.hiperglm}
\title{Print hiperglm Model}
\usage{
\method{print}{hiperglm}(object, ...)
}
\arguments{
\item{object}{A hiperglm model object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A warning message stating that the function is not implemented.
}
\description{
Placeholder function for printing a hiperglm model.
}
