% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.hiperglm}
\alias{vcov.hiperglm}
\title{Variance-Covariance Matrix for hiper_glm Objects}
\usage{
\method{vcov}{hiperglm}(object, ...)
}
\arguments{
\item{object}{An object of class "hiperglm".}

\item{...}{Further arguments passed to or from other methods (not used here).}
}
\value{
A matrix representing the variance-covariance of the estimated coefficients.
}
\description{
Computes the variance-covariance matrix of the estimated coefficients.
For linear models, the matrix is given by \eqn{\hat{\sigma}^2 (X^\top X)^{-1}},
with \eqn{\hat{\sigma}^2 = \frac{n}{n-p} \| y - X\hat{\beta} \|^2}.
For logistic models, the observed Fisher information is \eqn{X^\top W X},
where \eqn{W} is a diagonal matrix with entries \eqn{p_i(1-p_i)}.
The inverse is computed from the "R" factor of the QR decomposition of \eqn{X} (linear)
or \eqn{\sqrt{W}X} (logistic).
}
