% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.hiperglm}
\alias{vcov.hiperglm}
\title{Extract Variance-Covariance Matrix from hiperglm Model}
\usage{
\method{vcov}{hiperglm}(object, ...)
}
\arguments{
\item{object}{A hiperglm model object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A warning message stating that the function is not implemented.
}
\description{
Placeholder function for extracting the variance-covariance matrix from hiperglm model.
}
