% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.hiperglm}
\alias{vcov.hiperglm}
\title{Variance-Covariance Matrix for hiper_glm Objects}
\usage{
\method{vcov}{hiperglm}(object, ...)
}
\arguments{
\item{object}{An object of class "hiperglm".}

\item{...}{Further arguments passed to or from other methods (not used here).}
}
\value{
A matrix representing the variance-covariance of the estimated coefficients.
}
\description{
Variance-Covariance Matrix for hiper_glm Objects
}
